Consolidate Flask app into cogent package
