+++ cogent/templates/all_graphs.html
