+++ cogent/templates/current_values.html
