+++ tests/test_current_values.py
@@ -2,6 +2,7 @@ from datetime import datetime, timezone
 
 from sqlalchemy import create_engine
 
+from cogent import create_app
 from cogent.base.model import (
     Base,
     House,
@@ -14,7 +15,6 @@ from cogent.base.model import (
     init_data,
     init_model,
 )
-from flaskapp import create_app
 
 
 def test_current_values(monkeypatch, tmp_path):
